feat: finalizando estrutura da API, adicionando rotas, middlewares e controllers
