+++ routes/cadastroRouter.js
@@ -1,0 +1,9 @@
+import { Router } from "express";
+import { postCadastro } from ".././controllers/cadastroController.js"
+import { validarCadastro } from "./../middlewares/cadastroUsuarioMiddleware.js"
+
+const cadastroRouter = Router();
+
+cadastroRouter.post('/cadastro', validarCadastro, postCadastro);
+
+export default cadastroRouter;
